login component moved to UI blocks
